+++ assets/javascript/source/searchwp-live-search.js
@@ -5,18 +5,20 @@
 		this.config = null;
 
 		// internal properties
-		this.input_el = element;        // the input element itself
-		this.results_id = null;         // the id attribute of the results wrapper for this search field
-		this.results_el = null;         // the results wrapper element itself
-        this.parent_el = null;          // allows results wrapper element to be injected into a custom parent element
-		this.results_showing = false;   // whether the results are showing
-		this.form_el = null;            // the search form element itself
-		this.timer = false;             // powers the delay check
-		this.last_string = '';          // the last search string submitted
-		this.spinner = null;            // the spinner
-		this.spinner_showing = false;   // whether the spinner is showing
-		this.has_results = false;       // whether results are showing
-		this.current_request = false;	// the current request in progress
+		this.input_el = element;        		// the input element itself
+		this.results_id = null;         		// the id attribute of the results wrapper for this search field
+		this.results_el = null;         		// the results wrapper element itself
+        this.parent_el = null;          		// allows results wrapper element to be injected into a custom parent element
+		this.results_showing = false;   		// whether the results are showing
+		this.form_el = null;            		// the search form element itself
+		this.timer = false;             		// powers the delay check
+		this.last_string = '';          		// the last search string submitted
+		this.spinner = null;            		// the spinner
+		this.spinner_showing = false;   		// whether the spinner is showing
+		this.has_results = false;       		// whether results are showing
+		this.current_request = false;	     	// the current request in progress
+		this.results_destroy_on_blur = true;	// destroy the results
+
 		// kick it off
 		this.init();
 	}
@@ -71,22 +73,17 @@
                 // if parent_el was specified, inject the results el into it instead of appending it to the body
                 var swpparentel = $input.data('swpparentel');
                 if (swpparentel) {
-
                     // specified as a data property on the html input.
                     this.parent_el = $(swpparentel);
                     this.parent_el.html(results_el_html);
-
                 } else if (this.config.parent_el) {
-
                     // specified by the config set in php
                     this.parent_el = $(this.config.parent_el);
                     this.parent_el.html(results_el_html);
 
                 } else {
-
                     // no parent, just append to the body
                     $('body').append($(results_el_html));
-
                 }
 
 				this.results_el = $('#'+this.results_id);
@@ -125,9 +122,11 @@
 				}).keyup($.proxy(this.maybe_search, this));
 
 				// destroy the results when input focus is lost
-				$('html').click(function(){
-					self.destroy_results();
-				});
+				if(this.config.results_destroy_on_blur){
+					$('html').click(function(){
+						self.destroy_results();
+					});
+				}
 				$input.click(function(e){
 					e.stopPropagation();
 				});
